:fire: Farewell InterestRateModule

+++ test/fuzz/LendingPool/SyncInterests.fuzz.t.sol
@@ -27,7 +27,7 @@ contract SyncInterests_LendingPool_Fuzz_Test is LendingPool_Fuzz_Test {
         uint24 deltaTimestamp,
         uint128 realisedDebt,
         uint120 realisedLiquidity,
-        uint256 interestRate
+        uint80 interestRate
     ) public {
         vm.assume(realisedDebt <= type(uint256).max / RiskConstants.RISK_FACTOR_UNIT); // No overflow Risk Module
         // Given: deltaTimestamp than 5 years, realisedDebt than 3402823669209384912995114146594816 and bigger than 0
